backend/requirements.txt
fastapi>=0.134.0
google-genai>=1.0.0
httpx>=0.28.0
openpyxl>=3.1.0
xlrd>=2.0.0
pandas>=2.0.0
pydantic>=2.12.5
python-dotenv>=1.2.1
python-multipart>=0.0.22
supabase>=2.28.0
uvicorn[standard]>=0.41.0
